linux/clgl: Don't force load libraries by name
In a CL/GL sharing application we can't go and load another GL library
explicitely, it might not be the one the user has used.

But we know that whatever the user wants to share with, it has to
be loaded already when he requests the context, so we can just
lookup the symbols in the global scope.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>